Support alternate form of CPP line directives.
GCC supports two forms of line directive:

    #line <linenum> <filename>

    # <linenum> <filename>

Only the first of these is documented in [0], though it is actually the second
of these that the C pre processor outputs. CAmkES' understanding of these
directives was previously based on observed directives from CPP and it could not
understand the first of these. This commit adds support for the first form as
well, as it may be in use by some other non-CPP pre processing tool.

  [0]: https://gcc.gnu.org/onlinedocs/cpp/Line-Control.html