Fix: don't count including a template twice as a cycle.
This commit fixes an issue where, when traversing the include graph of a
template, all previously seen templates would be counted as ancestors of the
current one, rather than only its transitive parents. Note that the traversal
algorithm is somewhat inefficient as we will repeatedly scan a template that is
included multiple times, but it is assumed this is not a common enough scenario
to be a bottleneck for now.

Partial fix for JIRA CAMKES-439