[Fix] Prevent out-of-bound index on dataport flush
When using cached hardware dataports, users could previously specify an
arbitrary size for the region of the dataport to flush, potentially
causing an out-of-bounds array index when searching for the cap to the
frame backing the specified region of the dataport. This change prevents
the out-of-bounds index and returns with an error when an invalid range
is specified.