Removed race condition in passive thread init
RPC interfaces that use a passive thread (ie. no scheduling context)
need a temporary scheduling context during initialization. Passive RPC
threads run on this temporary scheduling context to perform some
initialization tasks, then use NBSendRecv to tell the main component
thread they are finished initializing, and block on and endpoint,
waiting for RPC calls. When the main component thread is notified of a
passive interface thread completing its initialization, it unbinds its
scheduling context.

Prior to this change, the interface thread told the main thread its
initialization was complete by doing a non-blocking send on and
endpoint. It was possible that the main thread wouldn't be waiting on
the endpoint by the time the non-blocking send was attempted. If the
main thread started waiting on the endpoint after the send was
attempted, it would never receive the message and block forever.

This commit replaces the endpoint with a notification. Passive interface
threads perform a non-blocking send on a notification when they are
finished initializing. If the main thread isn't waiting on the
notification when the send is performed, when it eventually waits on the
notification it will immediately receive it.