Removed passive init of notification interface
Passive initialization of notification interfaces was used in the case
when a user specifies a notification interface to not have a scheduling
context. In such a case, the notification interface thread would be
passive, and during its initialization, a scheduling context would be
bound to the notification object that the thread waits on. In this case,
there is a 1-1 mapping between the thread and notification, thus binding
a scheduling context to the notification is equivalent to using an
active thread. Thus the passive initialization of notification
interfaces is unnecessary.