Use configuration to find sc in templates
Some templates need to know the value of the sc attribute of an instance
or interface. A custom function was added to the template context for
looking up a given entity's sc. This change uses the configuration
settings, and camkes name mangling to find the sc instead, removing the
need for these custom lookup functions.