Fix: account for the python-capdl sources in the compilation cache.
Prior to this commit, a developer had to disable the compilation cache when
modifying the python-capdl module as it would not notice changes to these
sources and (incorrectly) retrieve stale entries from the level A cache. This
commit modifies CAmkES' version computation to treat the python-capdl sources
as part of itself. Though not technically correct, this more or less
corresponds to the way in which it is used. The effect of this commit is that
developers can now leave the compilation cache (and accelerator) enabled even
while working on python-capdl changes.

Closes JIRA CAMKES-481