Replace "sc" attribute with "passive" attribute
The "passive" attribute is a boolean which replaces the string attribute
"sc". The only use of the "sc" attribute was to mark a component or
interface thread as passive (ie. no scheduling context). This can now be
done directly by setting the "passive" attribute to true. Its default
value is false.