Repair a static analysis test of the accelerator.
In later versions of Cppcheck (1.72+ in my observation) certain compiler
built-ins are not natively understood. The relevant built-in for our case is
`__builtin_expect`. The result is that code like the following is incorrectly
flagged as a memory leak:

    void *foo(void) {
        void *p = malloc(100);
        if (__builtin_expect(p == NULL, 1))
            return NULL;
        return p;
    }

This commit passes another command-line flag to Cppcheck to tell it to assume a
GNU toolchain. This repairs this unit test.