libsel4muslccamkes: Fix underflow in mmap.
Prior to this commit, a large `malloc` or `calloc` would trigger an underflow in
calculating how much of the morecore region to remove. This underflow would go
undetected and result in a corrupted heap and invalid pointer being returned.
This commit checks for such an underflow. Note that it is still not possible to
fail an allocation in this way because a subsequent write to `errno` triggers a
data fault.

JIRA: SELFOUR-498, CAMKES-509