Use ReplyRecv instead of SignalRecv with rt
This fixes a bug where under certain conditions, no message would be
sent in reply to an rpc call. This is because SignalRecv was used to
reply to the client, rather than ReplyRecv. In order to make ReplyRecv
work, the saved reply cap must be swapped back into the server thread's
reply cap slot.