Implemented passive interface init protocol
Passive interface threads are initialised by capdl with bound scheduling
contexts. This is to allow the interface thread to perform some
initialisation and start waiting. It is up to the glue code generated
for the interface to relinquish its scheduling context when it's
finished initialising. It does this by sending on a notification, whose
cptr was passed as an argument to the interface's "run" function.
Typically this is done using SignalRecv, so the passive thread can send
on the notification and start waiting for input in a single call.

When the main thread of the component instance receives a notification
sent by a passive interface thread, it unbinds its scheduling context.
It is expected that the passive interface thread will be blocked waiting
for for IPC  when its scheduling context is unbound. If this isn't the
case, the thread will stop running and have no way of starting again
without having a scheduling context explicitly bound. This situation
could happen if another thread sends on the endpoint which the passive
thread was blocked on, before the main component instance thread got
around to unbinding the scheduling context. This can be prevented by
ensuring clients of passive servers run at a lower priority than the
server.