Swapping reply cap back into tcb by passive thread
In some situations, an interface thread that has been seL4_Called must
protect its reply cap from being wiped out (e.g. by seL4_Waiting). This
is achieved by moving the reply cap into its cspace. When replying in
these situations, the thread seL4_Sends on the reply cap, then
seL4_Recvs to wait for the next RPC call as a separate syscall. If this
was done by a passive thread on the rt kernel, the call to seL4_Send
would return the borrowed scheduling context, causing the interface
thread to stop before getting to the seL4_Recv that would see it waiting
for the next RPC (and accompanying scheduling context donation), leaving
it dead in the water.

Replying to the caller and blocking on its endpoint must be done as a
single syscall. The only way to do this is using seL4_ReplyRecv, which
requires the reply cap being in the tcb's reply cap slot. The rt kernel
provides a seL4_CNodeSwapCaller invocation for swapping a saved reply
cap back into the tcb. This change uses this to swap protected reply
caps back into the tcb, before calling seL4_ReplyRecv to reply and block
in a single syscall.