libsel4camkes: Replace use of libutils' linked-list with an inline version.
The linked-list interface in libutils that we were using prior to this commit is
deprecated and only in use by CAmkES. This commit removes CAmkES' use of it to
pave the way for removing the libutils version. Note that, as prior to this
commit, this linked-list is only used when assertions are enabled.