libsel4muslccamkes: Fix: Return ENOSYS instead of EPERM when _newselect unconnected.
The support for `_newselect` in this library is designed to detect and connect
to a dataport supplied in the application it is being linked against. In the
case when the dataport does not exist, this function was returning -1 which was
being translated to the error `EPERM`. This commits corrects this to return
`ENOSYS` indicating the syscall is unavailable.