Remove support for malloc debugging.
The functionality that is removed in this commit allowed users to enable a
`malloc` debugger (hosted in libsel4debug) globally in their systems. This
commit removes it because, in my observation, users never want to debug heap
allocation at this granularity. In a system that is sufficiently complicated to
require programmatic debugging in this style, it is not desirable to globally
instrument `malloc`. Rather, instrumenting `malloc` for a specific component
instance alone is useful.

Following this commit, the onus is on the user to set the LDFLAGS for a specific
component instance to use this debugging functionality. This requirement also
hopefully makes it clearer to the user that they need to link this component
instance against libsel4debug, which is not one of the core CAmkES libraries.