libsel4muslccamkes: Remove unnecessary exported header.
This commit removes a header arch_stdio.h from this library. As far as I can
tell, this header was never expected to be included from sources outside this
library. Marking the function as `noinline` had no effect, as it is never
visible in the same translation unit, so cannot be inlined anyway. Similarly,
marking the function `weak` had no effect as the way in which it was called did
not take advantage of this. This commit removes both these attributes.