Always set morecore region, even when heap size is 0.
Following this commit, a morecore region is unconditionally configured when
CAmkES applications start up. The effect of this is that a confusing error is
avoided when applications without a heap attempt to `malloc`. It is still not
possible to allocate dynamic memory without a heap, but at least the error
triggered is more sensible now.