libsel4muslccamkes: Remove dynamic mmap functionality.
The code removed in this commit is support for interoperating with libsel4vspace
in the context of a dynamic address space. In CAmkES, all address spaces are
statically constructed so this code was not in use. There should be no
user-visible functional changes from this commit. It simply removes some dead
code.