libsel4camkes: Make set_tid_address always succeed.
This system call is not actually exposed by Musl C, so user code can never
invoke it. However, Musl C calls it internally when a thread starts. It also
takes advantage of a property of this system call - that it is never supposed
to fail - and ends up saving the value `-1` as every thread's ID. Following
this commit, the thread ID Musl C saves corresponds to the CAmkES notion of
thread index. The thread ID is not currently used for any interesting behaviour
in the subset of Musl C's functionality that we use, but this future proofs us
against further extensions.

This commit also has the side effect of removing some spurious debug messages on
system start.