Fixed race condition in passive interface init
Prior to this change, each passive interface thread had a notification
cap which it used to signal to the main instance thread that it was
ready to have its SC unbound. The main instance thread would wait on
each notification sequentially and unbind the appropriate thread when
its signal was received. It is assumed that passive interface threads
have a lower priority than their main instance thread.

Since all passive interface threads were initialised concurrently, but
the main thread was only waiting on one endpoint at a time, it's
possible that the first passive interface thread to complete
initialising won't be the one on whose notification the main instance
thread waits on first. In such a situation, the main instance thread
will remain blocked on a notification of a not-yet-initialised interface
thread, and the just-initialised interface thread will be blocked on its
endpoint waiting for messages from clients. If a client sends a message
to the passive interface, it will wake up and run on its own scheduling
context, which hasn't been unbound by the main instance thread as it's
still blocked. This is a problem as if the main instance thread gets
around to unbinding the interface thread's scheduling context while the
interface thread isn't blocked on an endpoint, the interface thread will
be left in an un-runnable state.

This commit changes passive interface threads such that they start
without a bound scheduling context. For each component instance, there
is a single scheduling context used to initialise passive interface
threads. The main instance thread binds the scheduling context to each
passive interface thread, one at a time, allowing them to start waiting
on their endpoint before unbinding it. This solves the problem as
now the main instance thread unbinds passive interface thread scheduling
contexts as soon as they start waiting on their endpoints, assuming
passive interface threads have lower priority than the main instance
thread.