Support for nested lists and dicts in attributes
The CAmkES parser accepts specifications with nested lists and dicts in
attributes, but the runner fails to hash such attributes. This is
because its hashing logic has special cases for hashing lists and
dicts containing only hashable items. This extends the runner's hashing
logic to recursively apply a hashing function on collections.

Related issue: CAMKES-530