Allow old style N-1 connection syntax
In previous versions of CAmkES, N-1 connections could be declared by
using multiple connection statements with interfaces in common. CAmkES
next removed support for this, replacing it with connection statements
that allow multiple interfaces on either side. This proved too
restrictive, as it made it impossible to connect both external and
internal interfaces to an interface exposed by a hierarchical component.

This change adds back limited support for the old syntax. An interface
may appear in multiple connections, provided that:
- it is on the same side (to/from) of each connection
- the connector type of each connection is the same
- it is the only interface on its side of each connection