Fix n-1 connection bug
The bug prevented n-1 connections being described as multiple
"connection" statements if the common end of the connection was on the
"from" side. The problem was caused by passing the same set of
connection ends to a function that collapses them into single
connections for "to" sides and "from" sides. The function modified the
connection ends set under certain conditions such that when it was
reused it no longer contained valid informaiton. The fix was to pass a
copy of the set to each function.