No longer reuse frames for large frame promotion
Previously when replacing contiguous small frames with large frames, one
of the small frames was transformed into the large frame, and the others
thrown away. This left the capdl object allocator in an inconsistent
state. After this change, all small frames are thrown away, and a shiny
new large frame is created to replace them.