Fix bug when deleting small frames
This function deletes paging structures backing an address space. It
starts with the lowest level of the paging hierarchy, and proceeds
upwards, on each iteration deleting parents of structures from the
previous iteration. One problem was that it never actually proceeded
upwards in the hierarchy. Another problem was that it wasn't able to
remove entries from the root structure. Another problem was that it
would attempt to delete already-deleted structures.