libsel4camkes: Renamed sock_data_data to sock_data
Some socket functions depend on the caller having a dataport interface
named "sock_data". In older version of camkes, dataport templates would
generate a variable named <interface>_data, and a separate variable
named <interface> which would be a pointer to the former. This is no
longer the case. This changes the socket functions to use the
<interface> variable (sock_data) instead.