Corrected underlining of template errors
The camkes runner likes to prepend a context string to the start of
error messages. Template error messages are two lines long, with the
second line containing an underline which must be appropriately aligned
to the first line. Prepending a string to the error message breaks this
alignment. The solution is for template errors to be a list of lines,
rather than a multiline string, and prepending the context to each line
in the runner.

JIRA: CAMKES-551