AST and Parse errors aren't cast to strings
To properly display multi-line errors messages, they must be passed to
the `die` function as a list or tuple of lines. AST and Parse errors
were being cast from their exception type to a string before passing to
die. Instead, we now pass a tuple of exception arguments, which will
contain each line as a separate tuple element.