Deterministically hash None when hashing AST
The value of hash(None) can be different across multiple invocations of
python. CAmkES caches depend on unchanged ASTs hashing to the same value
across multiple runs, and ASTs can contain fields set to None such that
the hash of None affects the hash of the ast. Thus, when a None field is
encountered when hashing an AST, a deterministic value is chosen as its
hash value (specifically, hash("None")).