CAMKES-629: Use old capDL fault endpoints for mcs
On MCS kernel fault ep caps that are passed to seL4_TCBConfigure refer
to the callers cspace rather than the target's cspace.  In the capDL
python library there are two ways for specifying a threads fault ep, the
old way passed the objects name and capDL will give it an object ID that
makes it possible for the capDL loader app to refer to the object at
runtime.  The other way is to just specify a CPtr in the target threads
cspace that capDL will blindly use, assuming that it refers to a correct
object.  This commit reverts to the first way for mcs kernels.