Refactored hardware dataport frame allocation
When using register_shared_variable only one caller can pass in already allocated frames.
This refactors the frame allocation so that in an N-to-1 scenario only the first template
passes in the frames to register_shared_variable, and the other templates receive the
frame caps through the per template data stash.