Preserve exact page structures in preallocated shared memory
Preallocated shared memory should not be large frame promoted, as the user had requested
precisely these frames be used. Previously there was a bug where the original frames
were kind of preserved, but without their backing paging structures, leading to a
corrupt system that could not be loaded.

This change ensures that if we want to preserve frames that we also preserve the
correct paging structures by not deleting them.