Remove outfile from RenderOptions
Passing 'outfile' as a render option is nonsensical for several reasons. Namely we
render *every* template, not just the template that is actually requested and which
will actually get written to 'outfile'. As a result it is simply incorrect for a
template to make any decisions based on 'outfile' as it has no way of knowing if
it is actually the 'correct' outfile for it or not. Due to this incorrectness we
shall prevent templates from performing nonsensical operations by removing outfile
from the options.