Allow multiple --item and --outfile pairs to be specified
Changes the single global item to be built with a set of items we want to build, and a
corresponding set of items that have been completed. The bulk of this is accomplished by
two things
 * Changing the `done` wrapper to complete the requested item and only exit the runner
   if all items have been completed
 * Wherever we check if a template generation step matches the current item we have to
   check if it matches any of the remaining items
Importantly this commit is not removing the checks that prevented multiple outfiles/items
when the caches are enabled as the caching and accelerator logic is not updated by this
commit to handle multiple items.