Take advantage of multi outfile support from the runner if possible
The CAmkES runner is able to emit multiple outfiles in a single invocation, provided no
caching is enabled. This changes the cmake build template to, if all caching is turned off,
generate single CAmkES invocations for as many templates as possible.

This is implemented by batching CAmkES generation requests until they are requested. Not all
CAmkES invocations can be batched into a single generation as some CAmkES steps require the
result of the compilation and linking of previously generated files.