cmake: Remove support for compilation cache and accelerator
This drops any support from the cmake build system for the compilation cache and
accompanying accelerator in favour of always using the multi outfile support
from CAmkES (which is incompatible with the cache).

Benchmarks show that builds without the cache using the multiple outfile support
perform as good or better than builds that use the cache. As a result this is the
first step to completely removing the cache.