cmake: Extend component instances
Allows for extending any given instance of a component with additional sources, includes,
cflags, libraries etc.

The motivating use case for this is to allow for a FileServer component that serves files
from an embedded CPIO archive. In this case if you want multiple instances that each serve
differenve files there must be a way to somehow configure their builds independently.

This provides a generic way to both support the FileServer example, as well as any other
kind of per component specialization that is desired.