cmake: Understand --elf argument for CAmkESGen
The --elf argument to the CAmkES tool is an 'optional' argument in that it only
needs to be given if requesting items that will process elf files. This allows
CAmkESGen to be used to request such items by also being able to give the
elf files that will be given to the final camkes invocation.