cmake: Only generate camkes-gen.cmake if it is out of date
The camkes-gen.cmake only needs to be recreated if either any of its dependencies have
changed, or if the invocation to the camkes tool itself has changed.

Doing this forces us to perform some manual dependency management which is highly
undesirable, but due to the extreme cost to invoking camkes this is worth it.

After this change it is now possible to re-invoke cmake and change cache variables in the
cmake-gui (that do not get passed to camkes) without the cost of needlessly regenerating the
camkes-gen.cmake file.