work around capDL mismatch
With theory generation enabled (e.g. CAMKES_ARCH_THY), the theories in
a different camkes “platform” (whatever this is) end up getting
different capDL specs compared to the actual system. Even worse, the
incorrect capDL is liable to stick in the obj_state cache and the
toolchain spits out a completely broken capDL loader.

Rather than trying to find the root cause of this issue, we work
around this by putting all templates in the "seL4" “platform”,
disregarding the semantics of what a platform is.