ast: Flexible type checking on connectors
Updated the Camkes connector to check on the optional attribute
'disable_interface_type_checking'.
Adding 'disable_interface_type_checking' to your
connector definition relaxes the type checking of the connector,
allowing multiple interface types over the one connection. Updated
the camkes ast representation of a connector to enable the
the checking of the attribute.