Rewrite seL4RPCCall to use an abstract 'connector'
This is a substantial refactor that is meant to make it easier to change how different
templated connections work without having to duplicate a template and change it. In
theory the seL4RPCCall template could be instantiated with a different 'connector' to
give it different behaviours, although the mechanism to change the used connector is not
done in this change.

Whilst intended to be a semantics preserving refactor there are some minor changes. Some
code gets trivial changes of being split into multiple statements or moved into different
blocks. The largest change is some error handling is changed to be 'less helpful' due to
difficulties in getting precise information to the point where the error code now has to be
generated and having nice abstractions.