Start reorganising camkes project structure
This adds a CMakeLists.txt file for building camkes libraries to support
projects using camkes without using the provided toplevel CMakeLists file camkes provides.
It is cleaner for a project to import all of the camkes helpers itself
if it wants to build a camkes project. This also allows the generated
camkes-gen.cmake file to be imported into the same CMake scope as the
component declaration calls, making it possible to declare source
files for Camkes components that have their own generation build rules.