camkes/parser: add query to camkes ADL
This commit adds a new feature called a query. A query appears in the
ADL on the RHS of a setting in the configuration section of the ADL, eg:

configuration {
    component.blah = query()
}

queries takes comma separated parameters that can be any literal (numeric_expr |
multi_string | list | dict).

Queries are parsed in a new parsing stage between stage7 and stage8.