Parser: Exceptions: Add exceptions specific the DTB binding syntax
The reason why these are necessary is that the test framework (testexamples.py)
checks for failures in the "bad" tests by looking for categories of exceptions.

In order to distinguish the DTB Binding exceptions from the other exceptions and
keep the tests clean, it was necessary to add a new set of exceptions which all derive
from a common base class.