Add an engine for resolving DTB Binding queries
This engine wraps around the pyfdt library and provides an suitable interface for
querying it by properties, node-names, node paths and aliases.

It forms the base for the DTB Binding feature. It currently supports pretty much
every bit of the syntax that is related to binding. Everything from *-indexing
to alias lookup, etc, etc.