templates: Require all dataports to be fixed size
It was possible for dataports to have an unknown size until the ELF file
had been built and for the size to be based on the type of the C object.
We now require the size of the dataport to be specified at template
rendering time. This enables better alignment optimisation and simpler
code for managing shared seL4 objects.