camkes: obtain capdl-python from PYTHONPATH
Prior to this change, capdl-python was passed in as an argument to the
rest script and dynamically installed in the environment. A downside is
that this means no camkes packages can be imported into the test runner
that need capdl, as it is not in the python path of the test runner.

This change makes the script expect capdl to be in the PYTHONPATH, for
future changes which require camkes modules to be imported.