camkes: rewrite alltests.py to use unit tests
Prior to this change, alltest.py ran indepedant test runners from each
package. This has several downsides: tests could easily be missed (if
the individual test runner did not include the correct classes) and
tests were parallelised poorly (only on a per package test run basis).

This change rewrites the script to use unittest dynamic test discovery
features and concurrentTestSuite, which allows tests to be discovered as
long as they are in a valid module considered by the script (modules
are listed alltests.py help) and tests can run highly concurrently (this
takes the test run time from 1 minute to 7s on my machine).

The script still allows subsets of the tests to be run by specifying the
list of packages to be tested. As before, all packages are tested if no
package is specified.