Filters: Remove two more filters
- set_tcb_caps: This filter sets a thread's VSpace and CSpace caps and
sets a threads CNode size based on a config value or how many slots have
been allocated.
- guard_cnode_caps: Sets the correct guard value on any Caps to CNodes.

As the CNode object supports updating Caps that have been registered to
it whenever finalise_size() is called on it, we can call this when
finalising the total spec, and the TCB caps can be set in the template
when the TCB is created.