templates: Remove Perspective use
In preparation for removing perspectives, we remove references to
Perspectives and encode the naming policies at the call sites. This does
introduce a few code smells where the naming convention is duplicated,
but these should slowly be removed and is necessary to allow the removal
of Perspectives which is likely a worthwhile tradeoff.