camkes: Remove Perspectives
Perspectives and NameMangling.py was mostly needed to enable the
Filters. As Filters are no longer used, Perspectives should be able to
be safely removed. One point of Perspectives were for encoding policy of
how configuration options were named for certain interfaces, but as new
options are added and new templates are added it seems to have been
easier to encode the naming policy directly in the template where the
config value is obtained.