templates: Refactor TCB creation for components
Create thread stacks, ipc buffers, scheduling contexts, and properties
in one place. Previously having this allocation occuring within a switch
statement made it very hard to understand and maintain.

Now macros.threads() returns an array of all the threads for a component
containing properties for each thread:
- name: The name used for creating objects
- interface: If the thread is an interface thread, what interface it is
for
- intra_index: Index of the thread within an interface that has more
than one thread. 0 if not an interface thread.
- stack_symbol: Name of the stack for this thread
- stack_size: Size of the stack
- ipc_symbol: Name of the ipc buffer symbol for this thread.

The template can then create resources for each thread based on the
contents of this array.