Filters: Remove set_tcb_info filter
This filter set the following properties on a TCB:
- elf: the name of the ELF file describing the threads address space
- ip: The program counter of the thread when it is started
- sp: The stack pointer of the thread when it is started
- addr: The address of the IPC buffer for the thread
- init: An array of arguments that are given to the thread.
- sc_slot: A cap to the scheduling context of the thread if it is not
'passive'

These properties are all mostly known when the TCB objects are created
in the templates and we don't need to set them in a filter. The only
special cases are the ip, sp, and addr values that need to be set based
on the virtual address of the symbol in the ELF file. We use simpleeval
to allow the template to save a query that calculates an offset of a
virtual address corresponding to a symbol name so that the values can be
easily calculated once the ELF files are built, but the policy for what
offset into the symbol (e.g. to allow for page guards) can still be
controlled by the template. This mechanism is not Camkes specific and
could be moved into the python-capdl-tool library at a future date.