cmake: Move config definitions to correct location
Define ones that are passed to the camkes tool in camkes.cmake and
define the ones that are used in autoconf.h in libsel4camkes. This is to
reduce the chances of the camkes tool handling a config option
differently from the code if it is present as both a cmdline option and
as an autoconfig.h variable.