camkes-gen.cmake: Rearrange calls to CAmkESGen
Rearrange the order files are declared to the camkes runner so that they
match the order the runner uses internally.

The camkes runner follows a particular order for rendering templates.
This order is important for historical reasons related to templates
having undeclared dependencies on other templates being evaluated before
or after them. In addition, the build rules at the time that invoked the
runner template by template didn't encode any dependency ordering. Given
that we can now represent the ordering in the build system, we can work
towards removing the ordering logic from the runner which should greatly
simplify it.