Call capdl_linker for final capdl spec generation
The Camkes tool is invoked in 3 or 4 phases (the extra last phase is for
integity refinement and is currently optional). The first 2 phases,
parsing the AST and then generating all of the templates are both very
camkes specific. The 3rd phase takes the ELF files for all components
and a record of all objects and caps currently allocated
(AllocatorState) and generates a capdl spec after constructing paging
objects described by the ELF files. This is pretty generic and is also
performed by non-camkes capDL projects, such as the tutorials.  Using
capdl_linker.py allows more capdl internals to be transparent to the
camkes runner and reduces its complexity while also allowing the
functionality to be used in other non-camkes applications.