Move object_sizes to capdl
This target creates a YAML file that is intended to be consumed
by the python-capdl-tool tool and as that tool is used by projects that
don't necessarily use camkes. Moving this to the capdl repo allows
projects to use capdl without having to also use camkes repo
dependencies. It also more clearly indicates that this fullfilling an
interface that the capdl tools define and require and isn't a camkes
specific implementation.