camkes-gen: Ninja depfile needs full file paths
The depfile created by a custom_command requires the target file to be a
relative path from CMAKE_BINARY_DIR but this is not known by the camkes
tool as it can be called from any directory. We rely on camkes
generating a file with full file paths that we then truncate the
CMAKE_BINARY_DIR off the front of.