templates: Fix a cap allocation bug in seL4DTBHardware connector
Previously the template allocates IRQ handler caps using the 'irq_%d'
key (%d is an index in a loop) in the 'alloc()' function. This presents
a problem when generating interrupts for other components in the same
assembly. This is due to the fact that the same IRQ handler is handed
out for all other components as the cap table notices that we are using
the same key for all the IRQ handlers.