parser: Default cpp option to set
The camkes parser currently ignores all lines starting with # as a
potential cpp directive left over from a cpp invocation. This can lead
to specs that were meant to have CPP applied are often valid even
without running CPP. This leads to some very confusing failures where
a build chokes on code that was supposed to be #ifdef'd out.