components: Clear out serial/timer.camkes for ARM
The top level serial and timers CAmkES file point to the CAmkES files
inside the arch folders. These arch serial and timers CAmkES file point
to the mach/ or plat/ folders. Thus to maintain compatibility with
existing components (TimeServer, SerialServer) with regards to the
removal of redundant components, we just clear out the ARM arch CAmkES
files to point to nothing.